move GpuClockCalibration from colltrace/ to utils/ and add a device handle (#2048)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/torchcomms/pull/2048

just moving around some files since they aren't specific to colltrace, and adding __CUDACC__ friendly bindings for our ring buffer apis

Reviewed By: siyengar

Differential Revision: D99166508
