+++ comms/utils/CMakeLists.txt
@@ -4,6 +4,7 @@ find_package(CUDA REQUIRED)
 
 file(GLOB_RECURSE UTILS_SOURCES
     "*.cc"
+    "*.cu"
 )
 file(GLOB_RECURSE UTILS_CU_SOURCES
     "*.cu"
@@ -21,6 +22,11 @@ add_library(utils OBJECT
 target_compile_features(utils PRIVATE cxx_std_20)
 target_compile_options(utils PRIVATE
     -fPIC
+)
+set_target_properties(utils PROPERTIES
+    CUDA_SEPARABLE_COMPILATION ON
+    CUDA_RESOLVE_DEVICE_SYMBOLS ON
+    POSITION_INDEPENDENT_CODE ON
 )
 target_include_directories(utils PUBLIC
     ${ROOT}
